Issue #11163: enforced filesize MissingJavadocTypeTags
